requirements.txt
click
dune-client
jsonlines
matplotlib
matplotlib_inline
numpy
pandas
python-dotenv
requests
seaborn
jupyter_client